Peter sprint 3

+++ MovieDatabase.Domain/Models/Studio.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Phone { get; set; }
+        public long Phone { get; set; }
         public string Email { get; set; }
         public string Website { get; set; }
         public string Address { get; set; }
